+++ src/CoreEx.Cosmos/CosmosDbContainerBase.cs
@@ -17,8 +17,10 @@ namespace CoreEx.Cosmos
     /// <typeparam name="TSelf">The <see cref="Type"/> itself.</typeparam>
     /// <param name="cosmosDb">The <see cref="ICosmosDb"/>.</param>
     /// <param name="containerId">The <see cref="Microsoft.Azure.Cosmos.Container"/> identifier.</param>
-    public abstract class CosmosDbContainerBase<T, TModel, TSelf>(ICosmosDb cosmosDb, string containerId) : ICosmosDbContainer<T, TModel> where T : class, IEntityKey, new() where TModel : class, IIdentifier, new() where TSelf : CosmosDbContainerBase<T, TModel, TSelf>
+    /// <param name="dbArgs">The optional <see cref="CosmosDbArgs"/>.</param>
+    public abstract class CosmosDbContainerBase<T, TModel, TSelf>(ICosmosDb cosmosDb, string containerId, CosmosDbArgs? dbArgs = null) : ICosmosDbContainer<T, TModel> where T : class, IEntityKey, new() where TModel : class, IIdentifier, new() where TSelf : CosmosDbContainerBase<T, TModel, TSelf>
     {
+        private CosmosDbArgs? _dbArgs = dbArgs;
         private Func<T, PartitionKey>? _partitionKey;
 
         /// <inheritdoc/>
@@ -28,15 +30,25 @@ namespace CoreEx.Cosmos
         public Container Container { get; } = cosmosDb.GetCosmosContainer(containerId);
 
         /// <summary>
-        /// Gets the <see cref="PartitionKey"/> from the <paramref name="value"/>.
+        /// Gets or sets the Container-specific <see cref="CosmosDbArgs"/>.
+        /// </summary>
+        /// <remarks>Defaults to <see cref="ICosmosDb.DbArgs"/> on first access.</remarks>
+        public CosmosDbArgs DbArgs
+        {
+            get => _dbArgs ??= new CosmosDbArgs(CosmosDb.DbArgs);
+            set => _dbArgs = value;
+        }
+
+        /// <summary>
+        /// Gets the <see cref="PartitionKey"/> from the <paramref name="value"/> (only <b>Create</b> and <b>Update</b> operations).
         /// </summary>
         /// <param name="value">The value to infer <see cref="PartitionKey"/> from.</param>
         /// <returns>The <see cref="PartitionKey"/>.</returns>
-        /// <exception cref="AuthorizationException">Will be thrown where the infered <see cref="PartitionKey"/> is not equal to <see cref="ICosmosDb.PartitionKey"/> (where not <c>null</c>).</exception>
+        /// <exception cref="AuthorizationException">Will be thrown where the infered <see cref="PartitionKey"/> is not equal to <see cref="DbArgs"/> (where not <c>null</c>).</exception>
         public PartitionKey GetPartitionKey(T value)
         {
-            var dbpk = CosmosDb.PartitionKey;
-            var pk = _partitionKey?.Invoke(value) ?? CosmosDb.PartitionKey ?? PartitionKey.None;
+            var dbpk = DbArgs.PartitionKey;
+            var pk = _partitionKey?.Invoke(value) ?? DbArgs.PartitionKey ?? PartitionKey.None;
             if (dbpk is not null && dbpk != PartitionKey.None && dbpk != pk)
                 throw new AuthorizationException();
 
@@ -44,10 +56,11 @@ namespace CoreEx.Cosmos
         }
 
         /// <summary>
-        /// Sets the function to determine the <see cref="PartitionKey"/>; used for <see cref="GetPartitionKey(T)"/>.
+        /// Sets the function to determine the <see cref="PartitionKey"/>; used for <see cref="GetPartitionKey(T)"/> (only <b>Create</b> and <b>Update</b> operations).
         /// </summary>
         /// <param name="partitionKey">The function to determine the <see cref="PartitionKey"/>.</param>
         /// <returns>The <typeparamref name="TSelf"/> instance to support fluent-style method-chaining.</returns>
+        /// <remarks>This is used where there is a value and the corresponding <see cref="PartitionKey"/> needs to be dynamically determined.</remarks>
         public TSelf UsePartitionKey(Func<T, PartitionKey> partitionKey)
         {
             _partitionKey = partitionKey;
@@ -91,13 +104,13 @@ namespace CoreEx.Cosmos
         /// <param name="id">The identifier.</param>
         /// <param name="cancellationToken">The <see cref="CancellationToken"/>.</param>
         /// <returns>The entity value where found; otherwise, <c>null</c> (see <see cref="CosmosDbArgs.NullOnNotFound"/>).</returns>
-        public Task<Result<T?>> GetWithResultAsync(object? id, CancellationToken cancellationToken = default) => GetWithResultAsync(id, new CosmosDbArgs(CosmosDb.DbArgs), cancellationToken);
+        public Task<Result<T?>> GetWithResultAsync(object? id, CancellationToken cancellationToken = default) => GetWithResultAsync(id, new CosmosDbArgs(DbArgs), cancellationToken);
 
         /// <summary>
         /// Gets the entity for the specified <paramref name="id"/>.
         /// </summary>
         /// <param name="id">The identifier.</param>
-        /// <param name="partitionKey">The <see cref="PartitionKey"/>. Defaults to <see cref="ICosmosDb.PartitionKey"/>.</param>
+        /// <param name="partitionKey">The <see cref="PartitionKey"/>. Defaults to <see cref="DbArgs"/>.</param>
         /// <param name="cancellationToken">The <see cref="CancellationToken"/>.</param>
         /// <returns>The entity value where found; otherwise, <c>null</c> (see <see cref="CosmosDbArgs.NullOnNotFound"/>).</returns>
         public async Task<T?> GetAsync(object? id, PartitionKey? partitionKey, CancellationToken cancellationToken = default) => await GetWithResultAsync(id, partitionKey, cancellationToken).ConfigureAwait(false);
@@ -106,10 +119,10 @@ namespace CoreEx.Cosmos
         /// Gets the entity for the specified <paramref name="id"/> with a <see cref="Result{T}"/>.
         /// </summary>
         /// <param name="id">The identifier.</param>
-        /// <param name="partitionKey">The <see cref="PartitionKey"/>. Defaults to <see cref="ICosmosDb.PartitionKey"/>.</param>
+        /// <param name="partitionKey">The <see cref="PartitionKey"/>. Defaults to <see cref="DbArgs"/>.</param>
         /// <param name="cancellationToken">The <see cref="CancellationToken"/>.</param>
         /// <returns>The entity value where found; otherwise, <c>null</c> (see <see cref="CosmosDbArgs.NullOnNotFound"/>).</returns>
-        public Task<Result<T?>> GetWithResultAsync(object? id, PartitionKey? partitionKey, CancellationToken cancellationToken = default) => GetWithResultAsync(id, new CosmosDbArgs(CosmosDb.DbArgs, partitionKey), cancellationToken);
+        public Task<Result<T?>> GetWithResultAsync(object? id, PartitionKey? partitionKey, CancellationToken cancellationToken = default) => GetWithResultAsync(id, new CosmosDbArgs(DbArgs, partitionKey), cancellationToken);
 
         /// <summary>
         /// Gets the entity for the specified <paramref name="id"/>.
@@ -143,7 +156,7 @@ namespace CoreEx.Cosmos
         /// <param name="value">The value to create.</param>
         /// <param name="cancellationToken">The <see cref="CancellationToken"/>.</param>
         /// <returns>The created value.</returns>
-        public Task<Result<T>> CreateWithResultAsync(T value, CancellationToken cancellationToken = default) => CreateWithResultAsync(value, new CosmosDbArgs(CosmosDb.DbArgs), cancellationToken);
+        public Task<Result<T>> CreateWithResultAsync(T value, CancellationToken cancellationToken = default) => CreateWithResultAsync(value, new CosmosDbArgs(DbArgs), cancellationToken);
 
         /// <summary>
         /// Creates the entity.
@@ -177,7 +190,7 @@ namespace CoreEx.Cosmos
         /// <param name="value">The value to update.</param>
         /// <param name="cancellationToken">The <see cref="CancellationToken"/>.</param>
         /// <returns>The updated value.</returns>
-        public Task<Result<T>> UpdateWithResultAsync(T value, CancellationToken cancellationToken = default) => UpdateWithResultAsync(value, new CosmosDbArgs(CosmosDb.DbArgs), cancellationToken);
+        public Task<Result<T>> UpdateWithResultAsync(T value, CancellationToken cancellationToken = default) => UpdateWithResultAsync(value, new CosmosDbArgs(DbArgs), cancellationToken);
 
         /// <summary>
         /// Updates the entity.
@@ -211,13 +224,13 @@ namespace CoreEx.Cosmos
         /// <param name="id">The identifier.</param>
         /// <param name="cancellationToken">The <see cref="CancellationToken"/>.</param>
         /// <returns>The entity value where found; otherwise, <c>null</c> (see <see cref="CosmosDbArgs.NullOnNotFound"/>).</returns>
-        public Task<Result> DeleteWithResultAsync(object? id, CancellationToken cancellationToken = default) => DeleteWithResultAsync(id, new CosmosDbArgs(CosmosDb.DbArgs), cancellationToken);
+        public Task<Result> DeleteWithResultAsync(object? id, CancellationToken cancellationToken = default) => DeleteWithResultAsync(id, new CosmosDbArgs(DbArgs), cancellationToken);
 
         /// <summary>
         /// Deletes the entity for the specified <paramref name="id"/>.
         /// </summary>
         /// <param name="id">The identifier.</param>
-        /// <param name="partitionKey">The <see cref="PartitionKey"/>. Defaults to <see cref="ICosmosDb.PartitionKey"/>.</param>
+        /// <param name="partitionKey">The <see cref="PartitionKey"/>. Defaults to <see cref="DbArgs"/>.</param>
         /// <param name="cancellationToken">The <see cref="CancellationToken"/>.</param>
         /// <returns>The entity value where found; otherwise, <c>null</c> (see <see cref="CosmosDbArgs.NullOnNotFound"/>).</returns>
         public async Task DeleteAsync(object? id, PartitionKey? partitionKey, CancellationToken cancellationToken = default) => (await DeleteWithResultAsync(id, partitionKey, cancellationToken).ConfigureAwait(false)).ThrowOnError();
@@ -226,10 +239,10 @@ namespace CoreEx.Cosmos
         /// Deletes the entity for the specified <paramref name="id"/> with a <see cref="Result"/>.
         /// </summary>
         /// <param name="id">The identifier.</param>
-        /// <param name="partitionKey">The <see cref="PartitionKey"/>. Defaults to <see cref="ICosmosDb.PartitionKey"/>.</param>
+        /// <param name="partitionKey">The <see cref="PartitionKey"/>. Defaults to <see cref="DbArgs"/>.</param>
         /// <param name="cancellationToken">The <see cref="CancellationToken"/>.</param>
         /// <returns>The entity value where found; otherwise, <c>null</c> (see <see cref="CosmosDbArgs.NullOnNotFound"/>).</returns>
-        public Task<Result> DeleteWithResultAsync(object? id, PartitionKey? partitionKey, CancellationToken cancellationToken = default) => DeleteWithResultAsync(id, new CosmosDbArgs(CosmosDb.DbArgs, partitionKey), cancellationToken);
+        public Task<Result> DeleteWithResultAsync(object? id, PartitionKey? partitionKey, CancellationToken cancellationToken = default) => DeleteWithResultAsync(id, new CosmosDbArgs(DbArgs, partitionKey), cancellationToken);
 
         /// <summary>
         /// Deletes the entity for the specified <paramref name="id"/>.